Add cherry_pick function
We've hit a number of issues with upstream patches merging and
breaking our cherry picks.  This adds a variation on the upstream
CI cherrypick function that will log failures instead of killing
the entire run.  This should still allow us to easily find old
cherry picks that can be removed, but makes the whole process less
error-prone.  It also means we don't have to worry about a cherry
pick working with LKG and not with master because it merged after
LKG was set.