Cherry picks that fail must fail the install
It is important that if cherry picks fail to apply cleanly the installation is
aborted. These aren't errors that we can ignore. If a patch fails to apply, it
should be fixed immediately. The current set actually leaves merge markers left
in the tripleo-image-elements due to a conflict. Instead of the install
continuing and trying to debug issues this will causer later on, the install
should fail and new commits should be pushed to correct/update the cherry
picks.