instack-deploy-overcloud increase timeout and make configurable
Currently no timeout is passed to heat, which means it will use the
default (one hour) timeout, which may not be enough if a large
deployment with lots of nodes is specified.

Increasing the timeout to 4 hours in this script aligns with the
token expiry for the undercloud keystone, so should be a sane
default.  Also adds a CLI option to enable this default to be
overridden when required.