Update instack-setup-host-rhel7 to use Delorean
Changes instack-setup-host-rhel7 to setup the needed repositories to use with a
midstream Delorean instance. Also adds a git clone of the needed
tripleo-puppet-elements repository. Although this repository is not initially
needed, the work to install the Undercloud via Puppet and deploy the Overcloud
via Puppet will rely on it, so go ahead and add it. Note that
tripleo-puppet-elements is not yet packaged, which is why need to git clone.

- deletes the previous no longer relevant instack-setup-delorean script.
- Update rhel7rc to set $ELEMENTS_PATH to include tripleo-puppet-elements.
- Also updates $DIB_YUM_REPO_CONF and the cloud image to use RHEL 7.1.
- Remove os-cloud-config element
  This element is no longer needed as there one is available in
  tripleo-image-elements.
- Updates json file for RHEL 7
  - tuskar-ui is removed b/c of it's required role specificity
  - we must run the finalise hook b/c use-ephemeral now has a required
    hook script that runs in finalise
  - exclude pip-and-virtualenv
  - The finalise hook for use-ephemeral is not very well idempotent so, also
    add some additional steps to 00-cleanup before so we can rerun
    installations.
- Set additional install types to package

Change-Id: Ia0e365b15b4b4620c5a392c2a5e3e13d1683b5ef