Use current delorean rdo-management repo
Things appear to be stable enough that we can use the current symlink'd
delorean rdo-management repo. I think this is more in line with what
people expect.

It also enables us to switch over to installing instack-undercloud from
packages since the current repo will be enabled.

Once the infrastructure is in place for a stable symlink'd repo, we'll
probably want to use that.

Change-Id: I7fd564f09f88a14a12a49634af84717ab032c698