Add CentOS 7 Support
Add support for running instack-install-undercloud on CentOS 7. The
selinux element from tripleo-image-elements fails when applying the
additional rabbitmq policy (depedencies are not met), so we add that
element to exclude_elements, and also inslucde selinux-permissive and
selinux-setenforce-0.

instack-setup-host-rhel7 is symlinked to an updated instack-setup-host and
supports both RHEL 7 and CentOS 7. We may like to consider removing the
rhos-release logic out of that script entirely and just document that
bit since rhos-release is solely for internal consumption.

Change-Id: I89a43f68632d8526b049be2b666c2426a5ccf603