Revert "Ironic: Add support for local boot upstream"
This reverts commit 6ffd942fae504b2e6e2cb4f4b2bba1317d8751ce.

The code to set the capabilities key in the node needs to come after the
discovery process and update the capabilities key if it already exists,
otherwise we blow away the matched profile.

There is some question as to whether there is a fix for ironic-discoverd
to update capabilities instead of replace, but I'm not sure what the
status of that is or if it's built in rdo-management.

Since this issue is affecting people right now, going to revert for the
time being.