Improve virt setup
Improve the virt setup by adding manual steps into instack-virt-setup.
The reason this wasn't done initially was that the flow to get the user
added to the libvirtd group is troublesome to automate. To work around
that, just use sudo throughout instack-virt-setup whereever we interact
with libvirt.

The user account is still added to libvirtd, we just leave it up to the
human user to choose when to relogin if they want to take advantage of
that.

Update documenation as well to remove the previous manual and
troublesome steps.

Change-Id: I61fe56cd01a1deb7abaa70262149022bb5f60085