Match 1 control and * compute profiles
When we get to a working overcloud deployment, we will be deploying 1
control node and the rest of the nodes will be compute. Update the
edeploy plugin state configuration file to match this expectation.

This will still work with the existing --deploy-nodes command b/c we'll
still end up with 3 nodes matched to compute after discovery.

Change-Id: Ibeec7a30b16fa88f4184974711d68dea8489020d