Do not register nodes in instack-deploy-overcloud
We have an entirely new registration and discovery process for
instack-undercloud. Given how that is supposed to work, it makes no sense to
register nodes in instack-deploy-overcloud any longer. This is typically a
no-op if the nodes are already registered. However, if they aren't, you end up
with registered nodes with no matching profile since discovery was not done.

Let's just remove the register nodes step so we fail earlier.

Change-Id: Ifd16598f9572b0755102c93d1d161ec717d5a259