Factor out overcloudrc logic so it can be used standalone
instack-test-overcloud has a bunch of logic for generating an
appropriate overcloudrc environment, but it's tricky to replicate
this manually.  Factor out those pieces into their own script that
generates a standalone overcloudrc file to make it easier to work
with the overcloud.  The standalone rc file is created so they
won't have to first source stackrc every time they want to work
with the overcloud.

Documentation of the new script is also added.

Change-Id: I3b8c99e9e013c27b21bd9a60b5a772dc3b5c1dd6