Fix call to rhos-release
A bug in rhos-release used to cause both 'poodle' and 'puddle'
repositories to be enabled when run with the -d option.

This bug was fixed, so now when you run with -d, you only get
the 'poodle' repository.  Without -d, you get the 'puddle'
repository.

The instack-setup-host-rhel7 script was calling it with '-d',
but then disabling the 'poodle' repository.  Prior to the bugfix
in rhos-release, this was equivalent to not running rhos-release
with '-d'.  With the new fix in rhos-release, this meant that
neither repository was enabled.

The fix is to simply call rhos-release without '-d' if the goal
is to have the 'puddle' repository enabled.

Change-Id: I334cfe86ad5e56c282e7e4184e25d78893f4b919
Signed-off-by: Lon Hohberger <lon@metamorphism.com>