Don't use rpm directly
This was originally done because we had no repos enabled and yum
chokes on that, but now the call is happening after delorean is
enabled so it isn't a problem.  Using yum allows you to set a proxy
in yum.conf, and yum just generally disapproves when you install
things without it.

Explicitly removes rhos-release if it's already installed so we
can install the latest version.

Change-Id: If6cedf1db5e6207696c5b6f0dcad97f1f6a58233