Re-use existing passwords
Check for the existence of tripleo-undercloud-passwords and reuse the
users/passwords/secrets from that file during instack-install-undercloud
if it exists.

This makes instack-install-undercloud re-runnable. Since user's are
already defined with passwords in Keystone, we need to reuse existing
ones on subsequent runs.

Additional users/passwords/secrets are written out to the
tripleo-undercloud-passwords file since this makes them much easier to
reread.

There is also no longer a need for the 00-cleanup script at all after
the migration to Puppet. This was part of the problem of
instack-install-undercloud not being re-runnable b/c the $OK file from
98-undercloud-setup was getting deleted so that script was re-running
causing all sorts of problems.

98-undercloud-setup itself has a small change where I experimented with
making it entirely re-runnable itself, but that got complex with the
call to setup-neutron, so I stopped going down that road. Still, the
small change is valuable to have.

Change-Id: I3cb3d8f695314840c2d77a2ec9fbea1ac9b8d52e