Ironic: Add support for local boot upstream
Now that local boot is supported upstream, this patch is updating the
scripts to support the upstream version of local boot.

This patch also depends on a new version of ironic-discoverd that
includes the commit cea076b4570f5f60ada9cae6eb7658b3ae4cc4fe, that fix
the problem of ironic-discoverd overwritten the node's capabilities if
they already exist.

Change-Id: I67d1b773373f9f6ededce74f19223dfe175d16a9