Increase mysql max_connections
On large undercloud systems, the default of 151 connections is not
sufficient because we start many of each api process, and each process
opens its own connection to the mysql server.  For example, on a
24 core system the overcloud deployment fails almost immediately
because the additional api traffic overwhelms the 151 connection
limit in only a few seconds.

On said 24 core system, 500 seemed to be more than enough.  During
deployment the number of connections seemed to level off a little
over 200.

Change-Id: I09c76883d456cfe2d578e2e2e18fe063db51dfba