Configure tuskar db
We've switched to puppet and are no longer using the boot-stack element
that was responsible for creating all the initial db's. However, tuskar
is still installed via the element, so we need to setup it's db
manually.

The needed configuration is added to config.json.template, including the
bootstrap dict since 90-tuskar-db-sync relies on os-is-boostrap-host.

Change-Id: Ifd3bc393f76918dd208fa1e80ef01cf973770625