Split out vendor-specific setup steps
We expect to eventually support multiple vendor-specific hardware
management tools, so we don't want to include all of this inline
in the deploy overcloud doc.  This creates a separate section,
linked from the deployment doc, that provides details on using
vendor-specific configuration tools.

Initially this is just populated with DRAC information because
that is the only vendor-specific functionality available.

Co-Authored-By: Imre Farkas <ifarkas@redhat.com>

Change-Id: I9103d310b15c9311fc2236dc84fe44de779dfaa7