Install python-tuskarclient
python-tuskarclient was not ending up installed once we removed
openstack-clients from the json files. That change along with the one to
switch to using instack-deploy-overcloud --tuskar must have raced
through CI at the same time, so the regression wasn't caught. The right
place to install python-tuskarclient is in the puppet manifest anyway.

Change-Id: I76945b84703277ad673d692a53934f3ff3c17efe