Add higher priority to root_device_hint plugin
Before RAID configuration is done, the standard plugin will raise an
error because of the missing volumes. The root_device_hint fakes the
scheduling value for the volume until it's fixed in later runs. That
means, during processing the root_device_hint plugin must be called
before the standard plugin.

Change-Id: I04ebcc391efd3410d0f00072bec3576ff2d7f90d