Fix number mixup in Get Images section
The indentation of the Get Images section is kind of all over the
place, and it breaks the numbering of the section.  Rather than
try to figure out which indentation is the problem, just make the
upload step a separate heading.  I think there's a strong argument
to be made that the upload step isn't part of the "get" process
anyway.

Change-Id: I6c6a21ed0b8c2a3e36137b1cfbf49f8f0023f21b