Discoverd Swift objects not accessible by the admin user
Swift objects containing the Ironic extra data are not
accessible by the admin user on the undercloud node.
We need the nodes data for the Cloudforms integration where
we are using the admin user to connect to the undercloud
services.

This patch gives the admin user the swiftoperator role on
the undercloud service tenant. This will allow CFME to
access this data by specifying the service tenant in API
requests.

Change-Id: I4e5bf4f130cd46e22156f75f52e6328bd1d62daa
Resolves: rhbz#1237064