Add wol ironic driver support
The basic pxe_wol driver in ironic allows ironic to be used with
even the most basic hardware setups. This patch allows instack-undercloud
to enable ironic support for the pxe_wol driver. Typically this
would only be utilised for home lab and poc setups, but significantly
expands the hardware that projects like TripleO can utilise.

Change-Id: Ifbfc1d079693d487e7409c1f4205410e801fc032