Decrease the IPMI retry timeout in Ironic
Upstream has a very conservative default for the value of
[ipmi] retry_timeout in ironic.conf. The downside of this
happens with slow or unresponsive BMCs. With the upstream
default of 60 seconds, the sync power state periodic task
can hang for 10 minutes with the node locked for the
duration.

The tradeoff is that older BMCs can crash if we set this
too low.

This patch lowers the default from 60s to 15s.

Change-Id: I5fe30d5978331c9e078341554c6b1008e11808ba