Add enable_tempest option
This patch adds an option to disable the installation of Tempest
packages in each undercloud. Users may or may not want
to run Tempest from their undercloud... not installing it
saves us network bandwidth and CI time upstream as well
(where we don't even have the extra time to execute Tempest
at this point currently anyway).

Change-Id: If18b3d333bdbdc20390cb8d952e8c6182218cc1d