Make 98-undercloud-setup rerunnable
This patch makes 98-undercloud-setup rerunnable, and drops the check for
the $OK_FILE so that the script is rerun every time "openstack
undercloud install" is executed.

This makes it such the installer can be rerun and the configuration from
undercloud.conf is always applied, without having to do things such as
manual db cleanup or hackery.

The neutron network and subnet are updated as long as the cidr is the
same.  Neutron does not allow updating a subnet cidr, so in the case
where the cidr is different, the network/subnet will be deleted and
recreated.

Note that as part of testing update the local_ip setting, I discovered
that bind_host for glance-registry was unset in the puppet hieradata,
causing glance registry to bind to 0.0.0.0. That has been fixed here as
well.

Change-Id: I7a8daa334ceb7f69fb28fe7cfc456aa1bf81e119