Correct ironic auth settings
The existing ironic.conf uses an unversioned keystone auth_url
setting. As the keystone API now defaults to v3 this can
cause provisioning failures if Ironic needs to communicate
with other services (neutron). This is because we haven't yet
updated TripleO to properly support configuration with
keystone V3.

We also need to set the ironic neutron_url because
if we don't puppet-ironic will use the auth_protocol and
auth_host parameters to construct it incorrectly if
we don't (since we no longer set those parameters).

Change-Id: Ief7fc3dbbd8b8c33ea3ec7bd127231698af27a7f
Depends-On: Ibcc63d82b6e7b9de6ef12209a43f39a4cdecb5a2
Closes-bug: #1495090