Disable Tuskar UI
We should not enable Tuskar UI by default. It's been broken due to
packaging issues for quite a while. While that appears to possibly be
fixed now by moving to a newer Delorean repo, we don't know for sure.
Best to disable it by default and let people opt-in, vs. forcing them to
opt-out.

Change-Id: I063c9f3db100a311fad8a518a1fb833527911ee7