Put staticweb middleware after keystoneauth in proxy
The staticweb middleware needs to be put after authentication
middlewares to ensure correct functionality as documented in
http://docs.openstack.org/developer/swift/middleware.html#staticweb

Without this Swift sends a HTML response even if the request was done
using a X-Auth-Token. This might result in a faulty handling of the
response on the client side; for example, "swift stat containername"
would report an empty, private container, while the container might
actually be public readable with data stored in it.

Change-Id: Id008847376914752d7a30ce7bdd8c33a44bfb4fe
Related-bug: 1494896