Remove tuskar from the undercloud install
There is currently no liberty tuskar package, so installs using
the latest liberty packages are failing when trying to do
tuskar dbsync. This patch resolves the issue by not installing
tuskar in the undercloud.

An alternative would be to fix the liberty packaging to build
in delorean(https://review.gerrithub.io/#/c/249035), but it
seems like we would not want to do this if we are not going
to support tuskar in liberty.

Change-Id: I06e7d399bc1dab167aae67ad335c758cd8dd5ec6