Install local docker registry with puppet
This is for overcloud usage for when containerized services are
deployed. Using a local registry allows for faster network transfers
after the initial pull to other nodes. The puppet config option to enable
this functionality is 'enable_docker_registry'.

The RPM dependencies for docker-registry not already included in instack
is as follows:
    python-blinker
    c-ares
    python-gunicorn
    libev
    python-gevent
    python-backports-lzma
    m2crypto

The registry default port of 5000 conflicts with keystone, so it has
been changed to 8787 instead. Iptables rule for port 8787 has been added
to allow outside access.

Change-Id: I40e871f9017315886b1aac256f3fd6b9778ac21b
(cherry picked from commit 623be791cea48583d50779cef1649508e912817f)