Make ceilometer hardware meters work out of the box
Although we were configuring the ceilometer hardware section on the
undercloud, none of the hardware meters were showing up in
ceilometer.  There are a couple of reasons for this:

1) The auto-generated readonly SNMP username was too long.  It seems
   to have a limit of 32 characters, and we default to generating
   a 40 character value.
2) The generated SNMP username wasn't automatically passed to the
   overcloud template, so by default the overcloud was deploying
   with an SNMP user named ro_snmp_user, regardless of what was
   configured on the undercloud.

To address these issues, this change sets the default SNMP user
to ro_snmp_user so it is not necessary to pass it as a parameter
in a default install.  It also makes a note of the length
restrictions on custom values.

Change-Id: I99768fd402a13056adef7fdafd8568ff5f08d1a1