Use flat_networks=* default for the ml2 plugin
We are setting our flat_networks setting for the ml2 plugin
manually to '' which as of neutron git commit
cb712c5a9ee2717f5d34b00a420529b842db90c7 causes a
'Flat networks are disabled' error to occur when
we attempt to create the ctlplane network.

This patch drops the configurion of flat_networks from
instack and instead relies on the puppet-neutron default
of '*' which will work.

Change-Id: I880c9a159a46d9a288cb1f8bf79af30e5045d5ab
Closes-bug: #1515315