Disable clean_nodes by default
This patch sets the conductor/clean_nodes configuration option
for Ironic to False by default. To maintain the same behavior
of the old ramdisk.

The tripleoclient `openstack baremetal introspection bulk`
command needs to be fixed to account for the time it takes
to move from manageable to available in order to have cleaning
turned on by default.

Change-Id: I23c3e7ada3da8085957e86406527e9e5202bdd69
Closes-Bug: 1517449