Disable auto activation of all volumes
A problem was discovered that occurs during the provisioning of an
Atomic host with ironic causing a kernel panic. The problem is that
during provisioning the thin-pool associated with the provisioned VM was
being auto-activated on the instack host.

With containerized deployments becoming more popular, this is an
important fix to get in now. The workaround of defining the
auto_activation_volume_list to an empty list should be unobtrusive to
any operations on the undercloud.

Closes-bug: #1521733

Change-Id: I0ebb6488b0c061b0aed00ad1560126126ef3777c
(cherry picked from commit c418409557607af245c7036df587686f8203bf75)