Set default Glance backend to Swift
This patch configures glance to use Swift as the default
backend for image storage. The 'file' backend is
still available (listed in known stores) so that users
who upgrade will still have access to those images.

The motivation for this change is to be able to load
images directly from Swift into the IPA deployment
ramdisk (via a swift temp URL). In the future...
say once we have support for "external swift"
in instack undercloud one could imagine using this feature
to offload most of the image deployment traffic from
the instack deployment vm/server entirely by relying
on an external swift cluster.

Change-Id: If57520cbc8ec2caa7a9cb3796419648318e5e414