Allow setting the hostname during undercloud install
Getting the hostname settings right is a particularly annoying and
tricky part of the undercloud install.  Allow users to specify a
fixed hostname in undercloud.conf, which we will then use to
automatically configure the hostname settings on the system.

Also improves the /etc/hosts check so it just adds the necessary
entries to /etc/hosts rather than raising an error when they are
missing.

Change-Id: Ice707153636c702be90c55adbbc5f081dcb0ee74