Disable the nova ec2 API
This will save us around 120MB of memory which is wasted
on this extra process in the undercloud.

NOTE: We could save more by running nova-api-os-compute and
nova-api-metadata as individual processes instead of using
the nova-api "launcher" (which wastes memory). It may be worth
just waiting for Apache mod_wsgi to try this out though.

Change-Id: I2d9ce88850bf821ac3bf8e4dfa35b2a5f2856fa8