Create default flavors as part of undercloud install
We mostly do node scheduling based on profile, which means that the
actual values in the flavors don't matter in many cases. To make
the entire deploy process easier, just create the default flavors
with some sane minimum specs.  They can always be changed later
for users who want to do something different.

As part of this change, the post deploy code in undercloud.py was
refactored to be a bit more modular and easier to test.

Depends-On: I330884ec207c8bcee31961a292499165931d660a
Change-Id: I5dec6dc59fb537718d5175e70fab1d9f273d686b