Set Nova's ram_allocation_ratio configuration option to "1.0" By default
the Nova's "ram_allocation_ratio" is "1.5", this doesn't make sense for
baremetal because we can't allocate more memory than the machine already
has.

By leaving it as default ("1.5") will cause scheduling instances to
fail in case the amount of RAM in the node's is exactly the same amount
requested in the flavor.

Closes-Bug: #1524752
Change-Id: I4e6cc3457c62ad1a0f3b81247e1718e492973e8c