Remove answers file support
This was a compatibility shim left in when this was converted from
bash to python, and it definitely should not be used anymore.  It
was only _ever_ used by a handful of people involved in the very
early days of instack-undercloud.

For reference, the answers file was replaced by undercloud.conf.

Change-Id: Id5e0136874623395d9c2b39fc863c67b20732406