Unset DIB_YUM_REPO_CONF before running undercloud install
This variable doesn't make sense in the context of an undercloud
install because we're not in a chroot so there's no reason to copy
repos in.  It also causes a confusing error message when it's set
because it tries to copy files over themselves.  For the sake of
sanity, let's just always unset it.

Change-Id: Ieb500d21828abcf56277ce056c1051d36bf4e460