Add ability to auto-generate self-signed certificates
Since it's a little tricky to get the certificate configuration
correct, and in most cases users of self-signed certificates are
going to be using them for testing purposes, let's add a way to
just auto-generate a certificate that will work correctly.

This is currently just codifying the documented process of
generating the certificate.  It checks for the existence of the
~/undercloud-[public_vip].pem file to maintain idempotency.  This
way the certificate will be re-generated if the VIP is changed, but
it won't be re-generated unnecessarily on runs where the VIP
remains the same.

Change-Id: I7616b82913d16f2ba837298309d836c4811c2a44