Add -w to iptables calls
If another process happens to be making an iptables call, the calls
in our post-configure scripts can be killed, with a message such as

Another app is currently holding the xtables lock. Perhaps you want
to use the -w option?

Adding -w will tell iptables to wait until the lock is available,
so such a situation won't kill the undercloud install.

Change-Id: Id7c9482c26954fd2dc86bc0ef3325bdd12dd8f42