Increase scheduler_max_attempts on undercloud and make it configurable
Interaction between Nova and Ironic is subject to races. It may happen that
Nova scheduler picks the same Ironic node for several instances. Of course only
one instance ends up deployed, the other fail with the Conflict exception. This
would repeat until all nodes are deployed, provided that we do have enough
Ironic nodes. However, with the current scheduler_max_attempts being equal to 3,
we fail long before.

The proper fix to this issue is being worked on in Nova and will probably merge
in Newton. Until it's done we work around this issue by increasing
scheduler_max_attempts substantially from 3 to 30. This number is essentially
random, and is based on the observation that people prefer to deploy nodes in
bulks of 20-30. This value also become configurable via the
scheduler_max_attempts configuration variable.

The only downside is that the actual scheduling problems will take longer now.
For TripleO we have additional node checks in the deployment command, so it
should not hit us too often.

Related: https://bugzilla.redhat.com/show_bug.cgi?id=1310178

Change-Id: I1f357cd0e1afe9e8ef6fd4cc7cb32c7d1a09583a