Set max_resources_per_stack to -1
The max_resources_per_stack configuration option for Heat engine is
meant to prevent DOS attacks by large stacks. For TripleO, the check
itself is actually quite resource instensive since it takes a while to
actually count the resources in all the nested stacks.

The tenant in the undercloud is trusted anyway.

Setting the option to -1 disables the check and makes a significant
improvement in overcloud deployment time.

Change-Id: Id5a36947d2dfa77fba9f90eb0e5be3eb27592696