Delete default flavors in Python
I noticed that it takes ~1 second to create all of our baremetal
flavors via Python, but around 15 to delete the default ones in bash.
I believe this is because of the startup overhead of the Nova CLI,
so doing the flavor operations in Python should cut some time off
the install.

Plus, less bash.  Everyone wins!

Change-Id: Id800f2153c4c8dca9ca19750c36bef22e57e45cc