Add an option to enable cleaning
Ironic supports automatic cleaning before and between deployments. With the IPA
image it wipes the hard drive, but may potentially do other tasks returning a
node to a clean state. We have it disabled by default, as it's time-consuming
and our bulk introspection operation does not account for it. However, there is
a valid use case for cleaning, so this patch adds an option to enable it.

Change-Id: Id4946ee08c67610aae4c28e5d28056105e2c0c84