Re-enable extra data collection during introspection
It was initially disabled for Liberty, as Liberty IPA images used to not
contain the required dependency (python-hardware-detect package).
However, after more discussions we decided that it's valuable to continue
having extra hardware data by default, as it used to be in Kilo
instack-undercloud.

The required package was in IPA since January.

Change-Id: Iccaefe72f2795dfb855a3b321c28328aef219d6b