Raise the mysql file limit
On boxes with many cores, the existing limit of 1024 could easily be
reached. Although we already pass open_files_limit as -1 to the puppet
class, this parameter actually has no effect, because mysqld is not
started as root when run under systemd, so it can't raise it's own open
file limit.

To get around this, create a systemd override file that sets the limit
to 16384.

It also removes the open_files_limit parameter to the mysql::server
puppet class, since it has no effect.

Change-Id: Id6cbdde370735186ee344c05ec722e19e319b059