Generate most of the pystache context automatically
For the most part, the pystache context is just a list of values from
undercloud.conf.  Rather than maintain this list manually, including
duplicate default values (some of which are bad, like the "unset"
password defaults), let's just generate the context dict based on
the options we can look up from undercloud.py.

This also avoids masking mistakes where a value is added to the
context, but a corresponding config opt is not added to
undercloud.conf, resulting in an unconfigurable "option".

Change-Id: I70cc0f98e4f869417e22270a71f1f18525a06aac