Force rebuild of ramdisk as part of overcloud-full
If we don't do this, then a number of modules will be missing from
the image's ramdisk.  One example is the multipath module - this
has a dependency on device-mapper-multipath, which is installed as
part of the image build.  However, by default the ramdisk is not
rebuilt after the package is installed so it never takes effect.

This is the particular case I'm trying to address because it
prevents deploying to diskless systems that use iscsi targets as
their root disks.  With this change, the necessary modules for such
a deployment will be in the ramdisk.

Change-Id: I3c6a94070ebba39896d00250a14b655969406459
(cherry picked from commit a4e59c7aeade6b1deaf919966bf712ba4cf5fab9)
(cherry picked from commit 0810f6f4f8bba41152eeb6d97b854bf847b7c717)