Temporarily set +e on systemd-journald restart for +bug/1564471
Though the restart of the systemd-journald service succeeds fine, an
assertion error during the restart of the systemd-journald service
causes the instack-undercloud install to fail (this is during an
undercloud upgrade in fact, details in the bug).

This review sets +e to allow the script to continue and then also
checks that the service was indeed started OK, exiting otherwise.

Related-Bug: 1564471
Change-Id: I38c14a6b143b51dec6f6f88f2e1cb591e07a08b3