Add hieradata override file
Adds a new configuration option for instack-undercloud,
hieradata_override. The option can be set to a path to a hieradata file
that can contain any custom hieradata.

When set, the file will be copied under /etc/puppet/hieradata and
configured as the first file in the hiera hierarchy in
/etc/puppet/hiera.yaml.

This allows for configuring any option for any service that is
configured via puppet (and exposes a way to configure options in its
module).

This should help with the increasing need to represent more and
more options in undercloud.conf.

Change-Id: If25ea0976d568d84317d8c70584150b56446ab64