Add a reload of keepalived
When SSL is enabled on the undercloud, during an upgrade, os-net-config
might restart network interfaces. This will cause keepalived to not
be able to keep the VIPs it controls up. The reload of the keepalived
gives it a chance to add the VIPs back where they should be.

Change-Id: Id4b1dbdd976bce0a157a16e1bd47033dbfca8675
Closes-Bug: #1570824
(cherry picked from commit e9119eace1a3afcc6d3b416639a302f0d14fea9a)