Use tripleo::selinux
Use the tripleo::selinux module to configure SELinux on the undercloud via
Puppet instead of the bash element.

For CentOS, SELinux is still permissive by defualt.
For RHEL, SELinux is still enforcing by defualt.

An additional hiera (CentOS.yaml/RedHat.yaml) file is added to the hierarchy to
capture this difference.

blueprint undercloud-elements
Change-Id: I0ee6648605fa10a2322edb8ec08514e5e0db1d02
Depends-On: If214005df733d41c2fa4e197df247d8a14baaa14