Don't use ironic-conductor element
This removes usage of the ironic-conductor element in favor of using
puppet instead.

However, puppet-ironic does not yet have full support for configuring
ipxe and tftp, so we still need to make use of 88-setup-ipxe and
69-ironic-tftp-support. Since those bash scripts need to run after
puppet has installed all the needed packages, they are moved to be
os-refresh-config scripts instead. This should be fine since those
scripts are idempotent.

blueprint undercloud-elements
Change-Id: I1a8c97e56309940e0a4394b7ed8c20e8e0c3c9a9