Don't use os-net-config element
Manage the installation of os-net-config and openvswitch and the
execution of os-net-config via puppet instead of the bash element.

We do still generate the os-net-config configuration via os-apply-config
so, a template is added for that to the undercloud-install element.

As part of this switch, we lose some of the logic from
20-os-net-config[1], namely the configure_safe_defaults fallback
function for when os-net-config failes. However, in almost all cases
(possibly 100%), we never want this fallback on the undercloud anyway
because it will reconfigure the management interface to the undercloud,
which presumably is the interface that is already in use for out of band
management.

This change also requires a 2 step puppet apply, with the first step
running os-net-config.

[1] https://github.com/openstack/tripleo-image-elements/blob/master/elements/os-net-config/os-refresh-config/configure.d/20-os-net-config

blueprint undercloud-elements
Change-Id: Ic19b45e138206ca88ccee5cf261a85c1bfab5e08