Deprecate default 192.0.2.0/24 CIDR
This is a non-routable CIDR per RFC 5737, so we shouldn't be using
it by default.  However, if we just change it and don't provide a
deprecation period we may break users who have deployed with the
default CIDR, so we should give them a cycle to update their
configs appropriately.

This changes the default of network_cidr to None so that we can
detect when a user has not changed our default, and sets the
appropriate override to maintain backwards compatibility in that
case.  It also prints a warning about the deprecation at the end
of the deploy process, when it is most likely to be noticed.

Change-Id: I931a1f7160a007d367621de5cc1034c56c7741cf
Partial-Bug: 1553222