Ensure systemd-journald is running
Puppet is used to create /var/log/journal, however, we weren't notifying
the systemd-journald service that it needed to be restarted after
creating the directory. This commit adds the notifcation, and ensures
the service is running.

Since the notification is added, we no longer need to express the
resource dependency between the directory and the service, so that line
is removed.

Change-Id: I4dfdf741fdf5ef1429429f27b4363bc7ec88c068