Add value format validations
There are circumstances where an opt may be set to a legal value
that will error in non-obvious ways.  Two of these are setting
local_ip to a non-CIDR-ish value which returns a cryptic netaddr
error, and using a non-FQDN for the undercloud hostname which will
cause failures later in the install.

This change adds validations to catch both of these cases.

Change-Id: Ied9b5ffaff18e2a405443a00b35ca53735c46024