Update Gemfile to work in gate
Puppet-lint v1.1.0 (the most recent release in the 1.x series) has a
crashing bug, so we were depending on puppet-lint HEAD. On 2016-06-22,
puppet-lint v2.0.0 was released, so puppet-lint HEAD is now in the v2
series. But puppet-lint_variable_contains_upcase depends on puppet-lint
v1, so puppet-lint HEAD no longer satisfies our dependencies.

This patch pins the version of puppet-lint to the commit immediately
before the v2.0.0 release, so we get the bug fix and satisfy
puppet-lint_variable_contains_upcase's dependencies.

Change-Id: I3eab1b883a3dc5ca5d02d4c71b97507d36d7f349
Closes-bug: #1597849