Modify LVM filter fix to blacklist iSCSI
A previous fix (c418409557607af245c7036df587686f8203bf75) was made to
disable auto activation on the instack host, which has turned out to
cause problems for standard LVM usage. Instead of disabling auto
activation, blacklist OpenStack iSCSI devices. This serves the purpose
of letting users be able to have their LVM volumes activate as expected
automatically, while also not exposing the host to nasty surprises when
Ironic is provisioning a guest.

For some reason, augeas wouldn't accept creating a list as before, so
the list is created manually in string form.

Change-Id: I6bb3fea5344fe8e230c1493ac6fb1f73cf70d371