Increase Mistral Task Size limit
When selecting environments to deploy from the UI, every environment
filename gets added individually to 'environments' key in the Mistral
environment. Apparently it means we hit the Mistral execution size
limit pretty quickly.

This isn't needed for the CLI at this point because it bunches all the
environment files together into one big "user-environments.yaml" file,
for backwards-compabitility reasons.

Note that this Mistral value seems to affect both input and output for
task execution, which is why this patch was initially created as a
prerequisite for Ia9d2312e942229b659cfc65d4c370874e87f44da.

Change-Id: I24453b6d83171860065b050f437602e1f655d8b8
Closes-Bug: #1640815