galera and rabbit do not need to be accessible over the network
Both rabbitmq and galera do not need to be accessible from any network
except from the undercloud. By dropping the corresponding firewall rules
we make sure that only the undercloud can access these services.
All openstack services use the 192.0.2.X ip address and not
localhost. This all keeps on working because we have the following
netfilter rule:
 1563 93780 ACCEPT     all  --  lo     *       0.0.0.0/0 0.0.0.0/0            /* 002 accept all to lo interface */ state NEW

The network stack is smart enough to realize that these are all local
packets. A cleaner approach might be to actually make sure
that rabbitmq and galera only listen to localhost and configure all
the openstack services to talk to localhost.

Note that in liberty and mitaka we did not let galera be reachable from
the network, so this is a newton regression likely caused by the move
to puppet managing the firewall.

Future work should make rabbit and galera listen to localhost by
default as they are not needed to be exposed to the network at all.

Closes-Bug: 1617537
Change-Id: Iba66c0f5127ed3d394f9482b5843ddaf80538a76