Add constraints needed for ironic and inspector
When trying to initiate ironic-inspector it immediately tries to
access the endpoint. This endpoint might be unavailable, depending if
ironic-inspector started before HAProxy. So we need to arrange the
ordering so HAProxy starts before that.

On the other hand, mistral-db-populate needs inspector to be running,
else it will provision the baremetal.introspection actions without
inputs.

Change-Id: I227df7b35db000169dba167773d1f0f7e29b1527