Use http for Ironic deployments
When conductor/api_url is not configured in ironic.conf, Ironic queries
keystone for the url. When using Undercloud ssl, this results in
deployments using ssl and IPA is not able to talk to Ironic over ssl
because it does not trust the certificate.

Previously, this was not the case, and it would choose the internal url
instead of the public url. But something has apparently changed
somewhere and the tripleo-ci promote jobs using undercloud ssl are now
failing to promote due to this issue.

To restore the previous behavior, this patch configures
conductor/api_url to use the internal endpoint.

Dependst-On: I558b53591b14ed43c725a4d0e0a67401adc7d2f0
Co-Authorized-By: James Slagle <jslagle@redhat.com>
Co-Authorized-By: Emilien Macchi <emilien@redhat.com>

Change-Id: Ib99b8a0bec3b8235a32dab4a67a448ec89707f8a
Closes-Bug: #1613088