(mitaka only) Fix sysctl_settings configuration
sysctl settings are already managed in
puppet-tripleo/manifests/loadbalancer.pp (stable/mitaka only) so we only
want to set them when service_certificate is set to undef (SSL disabled)
otherwise the loadbalancer class will do it for us.

This is a Mitaka patch only because in master we use composable services
and don't have this problem.

Change-Id: I6938ea5f80badde3cb75d64012a1187061fa7fd0